requirements.txt
streamlit
requests
boto3
numpy
python-dotenv
